requirements.txt
black==24.1.1
click==8.1.7
flask-cors==4.0.0
flask==3.0.2
psycopg2==2.9.9
pydantic==2.6.1
python-dotenv==1.0.1
sqlalchemy==2.0.25
